Remove `Expenses` namespace
The only thing in the `Expenses` namespace was the `TravelDistanceCalculator`
service. This has been renamed to `DistanceCalculatorService` and moved into
the top of the services directory.
